requirements.txt
django==3.2.13
django-allauth==0.50.0
python-dateutil==2.8.2
celery==5.2.6
redis==4.2.2
django-celery-results==2.3.1
djangorestframework==3.13.1
celery==5.2.6
django-extensions==3.1.5
channels==3.0.4
channels-redis==3.4.0

gunicorn==20.1.0
whitenoise==6.0.0